requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
entrypoints==0.3
flake8==3.7.7
future==0.17.1
idna==2.8
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dotenv==0.10.3
python-telegram-bot==11.1.0
redis==3.2.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
